@@ -5,7 +5,7 @@
 name: libclang_example
 
 environment:
-  sdk: '>=2.8.1 <3.0.0'
+  sdk: '>=2.12.0-29.10.beta <3.0.0'
 
 dev_dependencies:
   ffigen:
