@@ -16,7 +16,7 @@ final _logger = Logger('ffigen.header_parser.structdecl_parser');
 
 /// Holds temporary information regarding [struc] while parsing.
 class _ParsedStruc {
-  Struc struc;
+  Struc? struc;
   bool nestedStructMember = false;
   bool unimplementedMemberType = false;
   bool flexibleArrayMember = false;
@@ -29,11 +29,11 @@ class _ParsedStruc {
 final _stack = Stack<_ParsedStruc>();
 
 /// Parses a struct declaration.
-Struc parseStructDeclaration(
+Struc? parseStructDeclaration(
   Pointer<clang_types.CXCursor> cursor, {
 
   /// Optionally provide name (useful in case struct is inside a typedef).
-  String name,
+  String? name,
 
   /// Option to ignore struct filter (Useful in case of extracting structs
   /// when they are passed/returned by an included function.)
@@ -58,7 +58,7 @@ Struc parseStructDeclaration(
     );
     // Adding to seen here to stop recursion if a struct has itself as a
     // member, members are updated later.
-    bindingsIndex.addStructToSeen(structUsr, _stack.top.struc);
+    bindingsIndex.addStructToSeen(structUsr, _stack.top.struc!);
     _setStructMembers(cursor);
   }
 
@@ -66,7 +66,7 @@ Struc parseStructDeclaration(
     _stack.top.struc = bindingsIndex.getSeenStruct(structUsr);
 
     // If struct is seen, update it's name.
-    _stack.top.struc.name = config.structDecl.renameUsingConfig(structName);
+    _stack.top.struc!.name = config.structDecl.renameUsingConfig(structName);
   }
   return _stack.pop().struc;
 }
@@ -90,38 +90,38 @@ void _setStructMembers(Pointer<clang_types.CXCursor> cursor) {
     _logger.fine(
         '---- Removed Struct members, reason: struct has array members ${cursor.completeStringRepr()}');
     _logger.warning(
-        'Removed All Struct Members from: ${_stack.top.struc.name}(${_stack.top.struc.originalName}), Array members not supported');
-    return _stack.top.struc.members.clear();
+        'Removed All Struct Members from: ${_stack.top.struc!.name}(${_stack.top.struc!.originalName}), Array members not supported');
+    return _stack.top.struc!.members.clear();
   } else if (_stack.top.nestedStructMember) {
     _logger.fine(
         '---- Removed Struct members, reason: struct has struct members ${cursor.completeStringRepr()}');
     _logger.warning(
-        'Removed All Struct Members from ${_stack.top.struc.name}(${_stack.top.struc.originalName}), Nested Structures not supported.');
-    return _stack.top.struc.members.clear();
+        'Removed All Struct Members from ${_stack.top.struc!.name}(${_stack.top.struc!.originalName}), Nested Structures not supported.');
+    return _stack.top.struc!.members.clear();
   } else if (_stack.top.unimplementedMemberType) {
     _logger.fine(
         '---- Removed Struct members, reason: member with unimplementedtype ${cursor.completeStringRepr()}');
     _logger.warning(
-        'Removed All Struct Members from ${_stack.top.struc.name}(${_stack.top.struc.originalName}), struct member has an unsupported type.');
-    return _stack.top.struc.members.clear();
+        'Removed All Struct Members from ${_stack.top.struc!.name}(${_stack.top.struc!.originalName}), struct member has an unsupported type.');
+    return _stack.top.struc!.members.clear();
   } else if (_stack.top.flexibleArrayMember) {
     _logger.fine(
         '---- Removed Struct members, reason: incomplete array member ${cursor.completeStringRepr()}');
     _logger.warning(
-        'Removed All Struct Members from ${_stack.top.struc.name}(${_stack.top.struc.originalName}), Flexible array members not supported.');
-    return _stack.top.struc.members.clear();
+        'Removed All Struct Members from ${_stack.top.struc!.name}(${_stack.top.struc!.originalName}), Flexible array members not supported.');
+    return _stack.top.struc!.members.clear();
   } else if (_stack.top.bitFieldMember) {
     _logger.fine(
         '---- Removed Struct members, reason: bitfield members ${cursor.completeStringRepr()}');
     _logger.warning(
-        'Removed All Struct Members from ${_stack.top.struc.name}(${_stack.top.struc.originalName}), Bit Field members not supported.');
-    return _stack.top.struc.members.clear();
+        'Removed All Struct Members from ${_stack.top.struc!.name}(${_stack.top.struc!.originalName}), Bit Field members not supported.');
+    return _stack.top.struc!.members.clear();
   } else if (_stack.top.dartHandleMember) {
     _logger.fine(
         '---- Removed Struct members, reason: Dart_Handle member. ${cursor.completeStringRepr()}');
     _logger.warning(
-        'Removed All Struct Members from ${_stack.top.struc.name}(${_stack.top.struc.originalName}), Dart_Handle member not supported.');
-    return _stack.top.struc.members.clear();
+        'Removed All Struct Members from ${_stack.top.struc!.name}(${_stack.top.struc!.originalName}), Dart_Handle member not supported.');
+    return _stack.top.struc!.members.clear();
   }
 }
 
@@ -142,7 +142,7 @@ int _structMembersVisitor(Pointer<clang_types.CXCursor> cursor,
         _stack.top.nestedStructMember = true;
       } else if (mt.broadType == BroadType.ConstantArray) {
         _stack.top.arrayMember = true;
-        if (mt.child.broadType == BroadType.Struct) {
+        if (mt.child!.broadType == BroadType.Struct) {
           // Setting this flag will exclude adding members for this struct's
           // bindings.
           _stack.top.nestedStructMember = true;
@@ -161,7 +161,7 @@ int _structMembersVisitor(Pointer<clang_types.CXCursor> cursor,
         _stack.top.unimplementedMemberType = true;
       }
 
-      _stack.top.struc.members.add(
+      _stack.top.struc!.members.add(
         Member(
           dartDoc: getCursorDocComment(
             cursor,
@@ -169,7 +169,7 @@ int _structMembersVisitor(Pointer<clang_types.CXCursor> cursor,
           ),
           originalName: cursor.spelling(),
           name: config.structDecl.renameMemberUsingConfig(
-            _stack.top.struc.originalName,
+            _stack.top.struc!.originalName,
             cursor.spelling(),
           ),
           type: mt,
