@@ -10,8 +10,7 @@ import 'data.dart';
 bool shouldIncludeStruct(String usr, String name) {
   if (bindingsIndex.isSeenStruct(usr) || name == '') {
     return false;
-  } else if (config.structDecl == null ||
-      config.structDecl.shouldInclude(name)) {
+  } else if (config.structDecl.shouldInclude(name)) {
     return true;
   } else {
     return false;
@@ -21,8 +20,7 @@ bool shouldIncludeStruct(String usr, String name) {
 bool shouldIncludeFunc(String usr, String name) {
   if (bindingsIndex.isSeenFunc(usr) || name == '') {
     return false;
-  } else if (config.functionDecl == null ||
-      config.functionDecl.shouldInclude(name)) {
+  } else if (config.functionDecl.shouldInclude(name)) {
     return true;
   } else {
     return false;
@@ -32,8 +30,7 @@ bool shouldIncludeFunc(String usr, String name) {
 bool shouldIncludeEnumClass(String usr, String name) {
   if (bindingsIndex.isSeenEnumClass(usr) || name == '') {
     return false;
-  } else if (config.enumClassDecl == null ||
-      config.enumClassDecl.shouldInclude(name)) {
+  } else if (config.enumClassDecl.shouldInclude(name)) {
     return true;
   } else {
     return false;
@@ -43,8 +40,7 @@ bool shouldIncludeEnumClass(String usr, String name) {
 bool shouldIncludeUnnamedEnumConstant(String usr, String name) {
   if (bindingsIndex.isSeenUnnamedEnumConstant(usr) || name == '') {
     return false;
-  } else if (config.unnamedEnumConstants == null ||
-      config.unnamedEnumConstants.shouldInclude(name)) {
+  } else if (config.unnamedEnumConstants.shouldInclude(name)) {
     return true;
   } else {
     return false;
@@ -54,7 +50,7 @@ bool shouldIncludeUnnamedEnumConstant(String usr, String name) {
 bool shouldIncludeMacro(String usr, String name) {
   if (bindingsIndex.isSeenMacro(usr) || name == '') {
     return false;
-  } else if (config.macroDecl == null || config.macroDecl.shouldInclude(name)) {
+  } else if (config.macroDecl.shouldInclude(name)) {
     return true;
   } else {
     return false;
@@ -66,7 +62,7 @@ final _headerCache = <String, bool>{};
 
 /// True if a cursor should be included based on headers config, used on root
 /// declarations.
-bool shouldIncludeRootCursor(String sourceFile) {
+bool shouldIncludeRootCursor(String? sourceFile) {
   // Handle null in case of system headers or macros.
   if (sourceFile == null) {
     return false;
@@ -78,5 +74,5 @@ bool shouldIncludeRootCursor(String sourceFile) {
         config.headers.includeFilter.shouldInclude(sourceFile);
   }
 
-  return _headerCache[sourceFile];
+  return _headerCache[sourceFile]!;
 }
