@@ -13,7 +13,7 @@ import 'package:ffigen/src/strings.dart' as strings;
 
 import '../test_utils.dart';
 
-Library actual, expected;
+late Library actual, expected;
 
 void main() {
   group('dart_handle_test', () {
