@@ -4,6 +4,7 @@
 
 import 'dart:io';
 
+import 'package:collection/collection.dart' show IterableExtension;
 import 'package:ffigen/src/code_generator.dart';
 import 'package:logging/logging.dart';
 import 'package:path/path.dart' as path;
@@ -14,8 +15,7 @@ import 'package:test/test.dart';
 extension LibraryTestExt on Library {
   /// Get a [Binding]'s generated string with a given name.
   String getBindingAsString(String name) {
-    final b = bindings.firstWhere((element) => element.name == name,
-        orElse: () => null);
+    final b = bindings.firstWhereOrNull((element) => element.name == name);
     if (b == null) {
       throw NotFoundException("Binding '$name' not found.");
     } else {
@@ -25,8 +25,7 @@ extension LibraryTestExt on Library {
 
   /// Get a [Binding] with a given name.
   Binding getBinding(String name) {
-    final b = bindings.firstWhere((element) => element.name == name,
-        orElse: () => null);
+    final b = bindings.firstWhereOrNull((element) => element.name == name);
     if (b == null) {
       throw NotFoundException("Binding '$name' not found.");
     } else {
@@ -53,7 +52,7 @@ void matchLibraryWithExpected(
       file.delete();
     }
   } catch (e) {
-    print('Failed test: Debug generated file: ${file.absolute?.path}');
+    print('Failed test: Debug generated file: ${file.absolute.path}');
     rethrow;
   }
 }
